pyka-scripts/requirements.txt
Faker==30.8.2
kafka-python==2.0.2
python-dateutil==2.9.0.post0
six==1.16.0
typing_extensions==4.12.2
